splited out in components thinking"
