+++ src/components/Footer.jsx
@@ -1,0 +1,15 @@
+// Footer.js
+
+function Footer() {
+    return (
+        <footer className="py-6 md:py-8">
+            <div className="container mx-auto">
+                <p className="text-center text-xs text-[#6A6A6A] lg:text-sm">
+                    Copyright ©2024 | All rights reserved by Learn with Sumit
+                </p>
+            </div>
+        </footer>
+    );
+}
+
+export default Footer;
